requirements/base.txt
Django==5.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
python-decouple==3.8
django-redis==5.4.0
redis==5.0.8
django-ratelimit==4.1.0
Pillow==11.2.1
httpx==0.24.1
drf-spectacular==0.28.0
channels==4.1.0
channels-redis==4.2.0
celery==5.4.0
flower==2.0.1
daphne==4.1.2